refactor(client): extract useFinanceData hook + add mobile bottom nav
- Extracted all data-fetching state into useFinanceData hook (~160 lines)
- App.tsx reduced from 759 lines to ~280 lines (63% reduction)
- Hook encapsulates: fetchData, fetchAllTransactions, fetchInvestmentValue, fetchApiVersion
- Added mobile bottom navigation bar (fixed, visible on < lg screens)
- Desktop nav unchanged; mobile users now have thumb-friendly tab bar
- Added pb-16 lg:pb-0 to body to prevent content hidden behind mobile nav
